Use Krita for rendering KRA files if there are linked files. No dependency detection works yet.

+++ renderchan/contrib/krita.py
@@ -11,9 +11,12 @@ class RenderChanKritaModule(RenderChanModule):
     def __init__(self):
         RenderChanModule.__init__(self)
         if os.name == 'nt':
-            self.conf['binary']=os.path.join(os.path.dirname(__file__),"..\\..\\..\\zip\\bin\\unzip.exe")
+            self.conf['zip_binary']=os.path.join(os.path.dirname(__file__),"..\\..\\..\\zip\\bin\\unzip.exe")
+            self.conf['binary']=os.path.join(os.path.dirname(__file__),"..\\..\\..\\krita\\bin\\krita.exe")
         else:
-            self.conf['binary']="unzip"
+            self.conf['zip_binary']="unzip"
+            self.conf['binary']="krita"
+
         self.conf["packetSize"]=0
 
     def getInputFormats(self):
@@ -31,15 +34,29 @@ class RenderChanKritaModule(RenderChanModule):
         comp = 0.0
         updateCompletion(comp)
 
-        random_string = "%08d" % (random.randint(0,99999999))
-        tmpPath=outputPath+"."+random_string
-        mkdirs(tmpPath)
+        deps_count=0
+        for item in extraParams["dependencies"]:
+            if not item.endswith(".conf"):
+                deps_count+=1
+                break
 
-        commandline=[self.conf['binary'], "-j", "-d", tmpPath, filename, "mergedimage.png"]
-        subprocess.check_call(commandline)
+        if deps_count==0:
 
-        os.rename(os.path.join(tmpPath,"mergedimage.png"), outputPath)
+            random_string = "%08d" % (random.randint(0,99999999))
+            tmpPath=outputPath+"."+random_string
+            mkdirs(tmpPath)
 
-        shutil.rmtree(tmpPath)
+            commandline=[self.conf['zip_binary'], "-j", "-d", tmpPath, filename, "mergedimage.png"]
+            subprocess.check_call(commandline)
+
+            #TODO: Compress image?
+            os.rename(os.path.join(tmpPath,"mergedimage.png"), outputPath)
+
+            shutil.rmtree(tmpPath)
+
+        else:
+            #TODO: PNG transperency settings at ~/.kde/share/config/kritarc ? use KDEHOME env ?
+            commandline=[self.conf['binary'], "--export", filename, "--export-filename", outputPath]
+            subprocess.check_call(commandline)
 
         updateCompletion(1)
